+++ src/app/Components/ComponentsSamples/Pickers/DateTimePickerSamplesViewModel.cs
@@ -22,7 +22,7 @@ public class DateTimePickerSamplesViewModel : ViewModel
         MaximumDate = new DateTime(2023, 05, 16);
         MinimumDate = new DateTime(2023, 01, 01);
 
-        Test = new DateTime(2023, 01, 28);
+        Test = new DateTime();
         Test2 = new DateTime(2024, 01, 29);
         Test3 = new DateTime(2023, 10, 29);
     }