+++ src/library/DIPS.Mobile.UI/Components/Pickers/DateAndTimePicker/iOS/DateAndTimePickerHandler.cs
@@ -18,7 +18,7 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
     {
         return new DUIDatePicker
         {
-            Mode = UIDatePickerMode.DateAndTime, PreferredDatePickerStyle = UIDatePickerStyle.Compact, VirtualView = VirtualView
+            Mode = UIDatePickerMode.DateAndTime, PreferredDatePickerStyle = UIDatePickerStyle.Compact, DateTimePicker = VirtualView
         };
     }
     
@@ -26,7 +26,7 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
     {
         base.ConnectHandler(platformView);
 
-        platformView.ValueChanged += OnDateSelected;
+        platformView.ValueChanged += OnValueChanged;
         platformView.SetInLineLabelColors();
         
         platformView.EditingDidBegin += OnOpen;
@@ -37,6 +37,8 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
     private void OnOpen(object? sender, EventArgs e)
     {
         m_isOpen = true;
+        
+        OnDateSelected();
     }
     
     private void OnClose(object? sender, EventArgs e)
@@ -85,8 +87,20 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
         handler.PlatformView.SetHorizontalAlignment(dateAndTimePicker);
     }
 
-    private void OnDateSelected(object? sender, EventArgs e)
+    private void OnValueChanged(object? sender, EventArgs e)
     {
+        OnDateSelected();
+    }
+    
+    private void OnDateSelected()
+    {
+        if (VirtualView.IsNullable && VirtualView.SelectedDateTime == default)
+        {
+            VirtualView.SelectedDateTime = DateTime.Now;
+            VirtualView.SelectedDateTimeCommand?.Execute(null);
+            return;
+        }
+        
         VirtualView.SelectedDateTime = (DateTime)PlatformView.Date;
         VirtualView.SelectedDateTimeCommand?.Execute(null);
     }
@@ -98,7 +112,8 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
 
     private static partial void MapSelectedDate(DateAndTimePickerHandler handler, DateAndTimePicker dateAndTimePicker)
     {
-        handler.PlatformView.SetDate(dateAndTimePicker.SelectedDateTime.Value.ConvertDate(), true);
+        handler.PlatformView.SetDate(dateAndTimePicker.SelectedDateTime.ConvertDate(), true);
+        handler.PlatformView.UpdatePlaceholders();
     }
 
     protected override void DisconnectHandler(DUIDatePicker platformView)
@@ -106,7 +121,7 @@ public partial class DateAndTimePickerHandler : ViewHandler<DateAndTimePicker, D
         base.DisconnectHandler(platformView);
 
         platformView.DisposeLayer();
-        platformView.ValueChanged -= OnDateSelected;
+        platformView.ValueChanged -= OnValueChanged;
         platformView.EditingDidBegin -= OnOpen;
         platformView.EditingDidEnd -= OnClose;
         DUI.OnRemoveViewsLocatedOnTopOfPage -= TryClose;