+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePicker/Android/MaterialDatePickerFragment.cs
@@ -47,10 +47,13 @@ public class MaterialDatePickerFragment : Object, IMaterialDateTimePickerFragmen
 
     private void SetDatePickerSelection(MaterialDatePicker.Builder builder)
     {
-        if(m_datePicker.SelectedDate is null)
+        if (m_datePicker.IsDateTimeOrTimeSpanDefault)
+        {
+            builder.SetSelection(DateTime.Now.ToLong());
             return;
+        }
         
-        var date = m_datePicker.IgnoreLocalTime ? m_datePicker.SelectedDate.Value : m_datePicker.SelectedDate.Value.ToLocalTime();
+        var date = m_datePicker.IgnoreLocalTime ? m_datePicker.SelectedDate : m_datePicker.SelectedDate.ToLocalTime();
 
         //Java uses the unix epoch, so we have find the total milliseconds from the date people have picked and the UnixEpoch start.
         builder.SetSelection(date.ToLong());