+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePicker/DatePicker.cs
@@ -1,6 +1,5 @@
 using DIPS.Mobile.UI.Components.Pickers.DateAndTimePicker;
 using DIPS.Mobile.UI.Components.Pickers.Platforms;
-using DIPS.Mobile.UI.Resources.Colors;
 using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
 namespace DIPS.Mobile.UI.Components.Pickers.DatePicker
@@ -16,22 +15,22 @@ namespace DIPS.Mobile.UI.Components.Pickers.DatePicker
         {
             base.OnHandlerChanged();
             
-            if(SelectedDate is null)
-                return;
-            
             // SelectedDate should not be above maximum date
             if (MaximumDate != null && SelectedDate > MaximumDate)
             {
-                var kind = SelectedDate.Value.Kind;
+                var kind = SelectedDate.Kind;
                 SelectedDate = DateTime.SpecifyKind(MaximumDate.Value, kind);
             }
 
             // SelectedDate should not be below minimum date
             if (MinimumDate != null && SelectedDate < MinimumDate)
             {
-                var kind = SelectedDate.Value.Kind;
+                var kind = SelectedDate.Kind;
                 SelectedDate = DateTime.SpecifyKind(MinimumDate.Value, kind);
             }
         }
+
+        public bool IsNullable { get; set; }
+        public bool IsDateTimeOrTimeSpanDefault => IsNullable && SelectedDate == default;
     }
 }