+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePicker/iOS/DUIDatePicker.cs
@@ -1,9 +1,13 @@
+using CoreGraphics;
 using DIPS.Mobile.UI.API.Library;
 using DIPS.Mobile.UI.Components.Chips;
+using DIPS.Mobile.UI.Components.Pickers.Platforms;
 using DIPS.Mobile.UI.Extensions.iOS;
+using DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings;
 using DIPS.Mobile.UI.Resources.Styles;
 using DIPS.Mobile.UI.Resources.Styles.Chip;
 using Microsoft.Maui.Platform;
+using ObjCRuntime;
 using UIKit;
 using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
@@ -14,98 +18,97 @@ public class DUIDatePicker : UIDatePicker
     private UIView? m_dateOrTimeView;
     private UIView? m_timeView;
 
-    private UIView? m_dateOrTimeViewButton;
-    
     private Chip? m_dateOrTimePlaceholderChip;
-    private Chip m_dateOrTimeChip;
+    private Chip? m_timePlaceholderChip;
+    
+    private bool m_firstDraw = true;
 
 #nullable disable
-    public View VirtualView { get; set; }
-    #nullable enable
-    
-    public void Initialize()
+    public IDateTimePicker DateTimePicker { get; set; }
+#nullable enable
+
+    public override void Draw(CGRect rect)
     {
-        UpdateInLineLayerAttributes(); //Update attributes when its first drawn
-        
-        ValueChanged += OnValueChanged;
-        EditingDidBegin += OnOpen;
+        base.Draw(rect);
+
+        if (!m_firstDraw)
+            return;
+
+        m_firstDraw = false;
 
         FetchSubviews();
-        
-        CreateChip();
-        
-        if (VirtualView is DatePicker { SelectedDate: null })
+
+        if (DateTimePicker is View view)
+        {
+            view.WidthRequest = Mode switch
+            {
+                UIDatePickerMode.Date => 121,
+                UIDatePickerMode.Time => 70,
+                _ => view.WidthRequest
+            };
+        }
+
+        var firstSubView = Subviews.First();
+        firstSubView.Frame = new CGRect(0, 0, firstSubView.Frame.Width - 100, firstSubView.Frame.Height);
+
+        if (DateTimePicker.IsNullable)
         {
             CreatePlaceholders();
         }
+
+        if (DateTimePicker.IsDateTimeOrTimeSpanDefault)
+        {
+            AddPlaceholders();
+        }
+
+        SetStyle(); 
     }
 
-    private void CreateChip()
+    public void UpdatePlaceholders()
     {
-        m_dateOrTimeChip = new Chip();
-        var nativeView = m_dateOrTimeChip.ToPlatform(DUI.GetCurrentMauiContext!);
-        
-        AddSubview(nativeView);
-        NSLayoutConstraint.ActivateConstraints([
-            nativeView.LeadingAnchor.ConstraintEqualTo(m_dateOrTimeView.LeadingAnchor),
-            nativeView.TrailingAnchor.ConstraintEqualTo(m_dateOrTimeView.TrailingAnchor),
-        ]);
+        if (DateTimePicker.IsDateTimeOrTimeSpanDefault)
+        {
+            AddPlaceholders();
+        }
+        else
+        {
+            RemovePlaceholders();
+        }
     }
 
     private void FetchSubviews()
     {
         m_dateOrTimeView = Subviews.FirstOrDefault()?.Subviews.FirstOrDefault();
-        m_dateOrTimeViewButton = m_dateOrTimeView?.Subviews.FirstOrDefault();
         if (Mode == UIDatePickerMode.DateAndTime)
             m_timeView = Subviews.FirstOrDefault()?.Subviews.LastOrDefault();
     }
 
-    private void OnOpen(object? sender, EventArgs e)
-    {
-        RemovePlaceholders();
-    }
-
     private void CreatePlaceholders()
     {
-        switch (Mode)
+        m_dateOrTimePlaceholderChip = new Chip
         {
-            case UIDatePickerMode.Date:
-            case UIDatePickerMode.Time:
-                {
-                    if(m_dateOrTimePlaceholderChip is null && m_dateOrTimeView is not null)
-                    {
-                        m_dateOrTimePlaceholderChip = new Chip { Title = "Dato", Style = Styles.GetChipStyle(ChipStyle.EmptyInput)};
-                        AddPlaceholder();
-                    }
-                    
-                    break;
-                }
-            case UIDatePickerMode.DateAndTime:
-                {
-                    m_dateOrTimeView = Subviews.FirstOrDefault()?.Subviews.FirstOrDefault();
-                    m_timeView = Subviews.FirstOrDefault()?.Subviews.LastOrDefault();
-                    
-                    if(m_dateOrTimeView is not null && m_timeView is not null)
-                    {
-                        //Set the width of the date picker to the width of the date and time picker plus their spacing
-                        VirtualView.WidthRequest = m_dateOrTimeView.Frame.Width + m_timeView.Frame.Width + (m_timeView.Frame.X - m_dateOrTimeView.Frame.Width);
-                    }
-                }
-                break;
-            case UIDatePickerMode.CountDownTimer:
-                break;
-            default:
-                throw new ArgumentOutOfRangeException();
-        }
+            Title = Mode == UIDatePickerMode.Time ? DUILocalizedStrings.Time : DUILocalizedStrings.Date,
+            Style = Styles.GetChipStyle(ChipStyle.EmptyInput)
+        };
+
+        if (Mode != UIDatePickerMode.DateAndTime)
+            return;
+
+        m_timePlaceholderChip = new Chip
+        {
+            Title = DUILocalizedStrings.Time, 
+            Style = Styles.GetChipStyle(ChipStyle.EmptyInput)
+        };
     }
 
-    private void AddPlaceholder()
+    private void AddPlaceholders()
     {
+        UIView nativeView;
         if (m_dateOrTimePlaceholderChip is not null && m_dateOrTimeView is not null)
         {
-            var nativeView = m_dateOrTimePlaceholderChip.ToPlatform(DUI.GetCurrentMauiContext!);
+            nativeView = m_dateOrTimePlaceholderChip.ToPlatform(DUI.GetCurrentMauiContext!);
             AddSubview(nativeView);
-            
+
             NSLayoutConstraint.ActivateConstraints([
                 nativeView.LeadingAnchor.ConstraintEqualTo(m_dateOrTimeView.LeadingAnchor),
                 nativeView.BottomAnchor.ConstraintEqualTo(m_dateOrTimeView.BottomAnchor),
@@ -113,37 +116,69 @@ public class DUIDatePicker : UIDatePicker
                 nativeView.HeightAnchor.ConstraintEqualTo(m_dateOrTimeView.HeightAnchor)
             ]);
         }
-        
+
+        if (m_timePlaceholderChip is null || m_timeView is null)
+            return;
+
+        nativeView = m_timePlaceholderChip.ToPlatform(DUI.GetCurrentMauiContext!);
+        AddSubview(nativeView);
+
+        NSLayoutConstraint.ActivateConstraints([
+            nativeView.LeadingAnchor.ConstraintEqualTo(m_timeView.LeadingAnchor),
+            nativeView.BottomAnchor.ConstraintEqualTo(m_timeView.BottomAnchor),
+            nativeView.TrailingAnchor.ConstraintEqualTo(m_timeView.TrailingAnchor),
+            nativeView.HeightAnchor.ConstraintEqualTo(m_timeView.HeightAnchor)
+        ]);
     }
 
     private void RemovePlaceholders()
     {
         m_dateOrTimePlaceholderChip?.ToPlatform(DUI.GetCurrentMauiContext!).RemoveFromSuperview();
+        m_timePlaceholderChip?.ToPlatform(DUI.GetCurrentMauiContext!).RemoveFromSuperview();
     }
-    
-    private void UpdateInLineLayerAttributes()
+
+    private void SetStyle()
     {
-        if (PreferredDatePickerStyle == UIDatePickerStyle.Inline) return; //Changing these colors when the style is inline will change the entire in line date picker. Which messes up the colors.
-        
+        if (PreferredDatePickerStyle == UIDatePickerStyle.Inline)
+            return; //Changing these colors when the style is inline will change the entire in line date picker. Which messes up the colors.
+
         //Tested on iOS 15,16,17
-        
+        if (m_dateOrTimeView is null)
+            return;
+
         //In line date picker
-        var inlineDateViewLayer = this.Subviews.FirstOrDefault()?.Subviews.FirstOrDefault()?.Subviews
-            .FirstOrDefault();
-        SetDefaultLayerAttributes(inlineDateViewLayer);
-        
+        SetDefaultLayerAttributes(m_dateOrTimeView.FindChildView<UIButton>());
+
+        var systemBackgroundViewForDateOrTimeView = FindSystemBackgroundView(m_dateOrTimeView);
+        if (systemBackgroundViewForDateOrTimeView is not null)
+        {
+            // Removes the slightly black overlay
+            systemBackgroundViewForDateOrTimeView.Alpha = 0;
+        }
+
+        if (m_timeView is null)
+            return;
+
         //In line time picker
-        var inLineTimeView = this.Subviews.FirstOrDefault()?.Subviews.LastOrDefault();
-        SetDefaultLayerAttributes(inLineTimeView);
-        
+        SetDefaultLayerAttributes(m_timeView.FindChildView<UIButton>());
+
+        var systemBackgroundViewForTimeView = FindSystemBackgroundView(m_timeView);
+        if (systemBackgroundViewForTimeView is not null)
+        {
+            // Removes the slightly black overlay
+            systemBackgroundViewForTimeView.Alpha = 0;
+        }
+    }
+
+    private static UIView? FindSystemBackgroundView(UIView root)
+    {
+        return root.Subviews.FirstOrDefault()?.Subviews.FirstOrDefault()?.Subviews.FirstOrDefault();
     }
 
     private static void SetDefaultLayerAttributes(UIView? view)
     {
-        if (view == null)
-        {
+        if (view is null)
             return;
-        }
 
         var defaultColor = Colors.GetColor(ColorName.color_secondary_30);
         var defaultCornerRadius = Sizes.GetSize(SizeName.size_2);
@@ -154,21 +189,10 @@ public class DUIDatePicker : UIDatePicker
 
     public void DisposeLayer()
     {
-        this.ValueChanged -= OnValueChanged;
+        m_dateOrTimeView = null;
+        m_timeView = null;
+        m_dateOrTimePlaceholderChip = null;
+        m_timePlaceholderChip = null;
     }
 
-    private void OnValueChanged(object? sender, EventArgs e)
-    {
-        UpdateInLineLayerAttributes();
-        UpdateChipTitle();
-    }
-
-    private void UpdateChipTitle()
-    {
-        var dateLabel = m_dateOrTimeView?.FindChildView<UILabel>();
-        if (dateLabel is not null)
-        {
-            m_dateOrTimeChip.Title = dateLabel.Text ?? string.Empty;
-        }
-    }
 }