+++ src/library/DIPS.Mobile.UI/Components/Pickers/Platforms/IDateTimePicker.cs
@@ -1,5 +1,7 @@
 namespace DIPS.Mobile.UI.Components.Pickers.Platforms;
 
-public interface IDateTimePicker
+public interface IDateTimePicker : IView
 {
+    bool IsNullable { get; set; }
+    bool IsDateTimeOrTimeSpanDefault { get; }
 }