+++ src/library/DIPS.Mobile.UI/Components/Pickers/TimePicker/TimePicker.cs
@@ -1,5 +1,4 @@
 using DIPS.Mobile.UI.Components.Pickers.Platforms;
-using DIPS.Mobile.UI.Resources.Colors;
 using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
 namespace DIPS.Mobile.UI.Components.Pickers.TimePicker;
@@ -10,4 +9,7 @@ public partial class TimePicker : View, IDateTimePicker
     {
         BackgroundColor = Colors.GetColor(ColorName.color_secondary_30);
     }
+
+    public bool IsNullable { get; set; }
+    public bool IsDateTimeOrTimeSpanDefault => IsNullable && SelectedTime == default;
 }